requirements.txt
# Function dependencies, for example:
# package>=version
Flask==1.0.3
google-cloud-firestore==1.2.0
